api/requirements.txt
flask==3.1.0
python-keycloak==5.3.1
flask-cors
pyjwt